Allow pattern replacement to ignore literals
Summary:
This is necessary because sometimes the patterns found have literals
include tuple of ints kind of literals. This values shouldnt be used for
pattern matching since often they are based on consts derived from
example inputs.

THis is not exactly a safe thing to do in general so by default it is
turned off

Test Plan:
Subsequent diff adds a pattern that relies on this

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 6937fa0a27c34adf7a70a06f20e34eb36182561c
Pull Request resolved: https://github.com/pytorch/ao/pull/2519